% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dat_1_0_1.R
\docType{data}
\name{Dat_1_0_1}
\alias{Dat_1_0_1}
\title{One stage and not adaptive to baseline}
\format{
An object of class \code{data.frame} with 8000 rows and 2 columns.
}
\usage{
Dat_1_0_1
}
\description{
This is data for one stage and not adaptive to baseline with continuous primary outcome
}
\keyword{datasets}
