% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dat_1_1_0.R
\docType{data}
\name{Dat_1_1_0}
\alias{Dat_1_1_0}
\title{One stage and adaptive to baseline}
\format{
An object of class \code{data.frame} with 8000 rows and 3 columns.
}
\usage{
Dat_1_1_0
}
\description{
This is data for one stage and adaptive to baseline with binomial outcome
}
\keyword{datasets}
