% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dat_2_0_1.R
\docType{data}
\name{Dat_2_0_1}
\alias{Dat_2_0_1}
\title{Two stages and not adaptive to baseline}
\format{
An object of class \code{data.frame} with 8000 rows and 4 columns.
}
\usage{
Dat_2_0_1
}
\description{
This is data for two stages and not adaptive to baseline with continuous primary outcome
}
\keyword{datasets}
