% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dat_2_1_0.R
\docType{data}
\name{Dat_2_1_0}
\alias{Dat_2_1_0}
\title{Two stages and adaptive to baseline}
\format{
An object of class \code{data.frame} with 8000 rows and 5 columns.
}
\usage{
Dat_2_1_0
}
\description{
This is data for two stages and adaptive to baseline with binomial outcome
}
\keyword{datasets}
