% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dat_3_0_0.R
\docType{data}
\name{Dat_3_0_0}
\alias{Dat_3_0_0}
\title{Three stages and not adaptive to baseline}
\format{
An object of class \code{data.frame} with 8000 rows and 6 columns.
}
\usage{
Dat_3_0_0
}
\description{
This is data for three stages and not adaptive to baseline with binomial outcome
}
\keyword{datasets}
