% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.myclass.R
\name{summary.myclass}
\alias{summary.myclass}
\title{Summarize the dataframe got from the package}
\usage{
\method{summary}{myclass}(object, ...)
}
\arguments{
\item{object}{A dataframe to be summarized}

\item{...}{other arguments in summary generic function}
}
\value{
Descriptive table of the dataframe is returned
}
\description{
Summarize the result of sequential primary outcome in `seqmeans`
and the estimated strategy values in `atsmeans`
}
\examples{
ats_outcome=atsmeans(data=codiacs,conf=TRUE,
 alpha=0.05,digits = 2,pch=18,xlab="abc")
summary(ats_outcome)

}
